+++ custom_model_runner/CHANGELOG.md
@@ -4,9 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-#### [1.1.4rc1] - in progress
+#### [1.1.4rc3] - in progress
+##### Added
+- r_lang fit component, pipeline, and template
 ##### Changed
 - search custom.py recursively in the code dir
+- set rpy2 dependcy <= 3.2.7 to avoid pandas import error 
 
 ## [1.1.3] - 2020-07-17
 ### Added